Fix table of contents linking
haddocset currently rewrites,

    <a name="v:foo">foo</a>

To

    <a href="../$pkgname-1.2.3/SomeModule.html#v:foo" name="v:foo">foo</a>

This seems to mess with Dash's (at least in Dash 3) table of contents in the
bottom left part of the window. It expects to append the `name` to the URL so
it ends up with a double-encoded version. For example,

    criterion-1.1.0.0/Criterion.html#t:Benchmark%23t:Benchmark

The fix for this is simply to leave the `<a name>` tags alone.

Alternatively,
[this version of ToC support](https://kapeli.com/docsets#tableofcontents)
(which I assume is new) will have to be implemented.