Add support for pkgroot relative paths in conf files
pkgroot relative paths are paths prefixed with `${pkgroot}`, which is
resolved to the directory in which the package config directory resides.

OSX ghc installations uses this to allow for ghc to be installed as an
application bundle. Therefore support for this is necessary in order to
pull in documentation for base packages on OSX.